+++ src/Nothing.ts
@@ -1,5 +1,3 @@
-import { Const } from "./const";
-
 export type Nothing = undefined | null;
 
 export type MNothing<T> = T | Nothing;
@@ -14,9 +12,3 @@ export const isNothing = (v: unknown): v is Nothing => v === null || v === undef
  * Return the default value if the provided value is Nothing
  */
 export const orElse = <T>(orElse: T, v: MNothing<T>): T => (isNothing(v) ? orElse : v);
-
-/**
- * Return the result of callable if value is Nothing
- * Use when the left value may require some computation, otherwise use `orElse`
- */
-export const onNothing = <A>(orCall: Const<A>) => (v: MNothing<A>): A => (isNothing(v) ? orCall() : v);
