--- src/const.ts
@@ -1,6 +1,0 @@
-export type Const<T> = () => T;
-
-/**
- * Wrap a vale in a function
- */
-export const _const = <A>(v: A): Const<A> => () => v;
