--- src/id.ts
@@ -1,3 +1,0 @@
-export const identity = <A>(a: A): A => a;
-
-export const id = identity;
