+++ src/index.ts
@@ -1,10 +1,2 @@
-export * from "./compose";
-export * from "./const";
-export * from "./curry2";
-export * from "./id";
-export * from "./lift";
-export * from "./mCall";
-export * from "./mApply";
-export * from "./mCompose";
-export * from "./mFn";
+export * from "./liftA2";
 export * from "./Nothing";
