+++ src/internals/utils.ts
@@ -16,3 +16,5 @@ export const prettify = (v: unknown): string => {
 export const nothings = [null, undefined];
 export const empties = [NaN, 0, "", [], {}];
 export const values = [1, "value", [1, 2], { a: 1 }, () => true, () => false];
+
+export const id = <T>(t: T): T => t;
